move Image and ImagePosition types in DescriptionPanel file
